# backend/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_files_jobs_upload_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_upload_files_jobs_upload_post",
        "type": "object"
      },
      "CreateBatchJobRequest": {
        "properties": {
          "items": {
            "description": "Absolute or relative local file paths",
            "items": {
              "type": "string"
            },
            "title": "Items",
            "type": "array"
          },
          "options": {
            "$ref": "#/components/schemas/ProcessingOptions"
          }
        },
        "required": [
          "options"
        ],
        "title": "CreateBatchJobRequest",
        "type": "object"
      },
      "CreateJobRequest": {
        "properties": {
          "mask_hints": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Mask Hints",
            "type": "object"
          },
          "options": {
            "$ref": "#/components/schemas/ProcessingOptions"
          },
          "upload_ids": {
            "items": {
              "type": "string"
            },
            "title": "Upload Ids",
            "type": "array"
          }
        },
        "required": [
          "options"
        ],
        "title": "CreateJobRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestPathsRequest": {
        "properties": {
          "paths": {
            "description": "Absolute local file paths",
            "items": {
              "type": "string"
            },
            "title": "Paths",
            "type": "array"
          }
        },
        "title": "IngestPathsRequest",
        "type": "object"
      },
      "JobFileResult": {
        "properties": {
          "engine_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Used"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "input_path": {
            "title": "Input Path",
            "type": "string"
          },
          "output_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Filename"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          },
          "provider_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Used"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "enum": [
              "queued",
              "processing",
              "done",
              "failed",
              "canceled"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "input_path",
          "state"
        ],
        "title": "JobFileResult",
        "type": "object"
      },
      "JobResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/JobFileResult"
            },
            "title": "Files",
            "type": "array"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/ProcessingOptions"
          },
          "progress": {
            "title": "Progress",
            "type": "number"
          },
          "state": {
            "enum": [
              "queued",
              "processing",
              "done",
              "failed",
              "canceled"
            ],
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "state",
          "progress",
          "created_at",
          "updated_at",
          "options"
        ],
        "title": "JobResponse",
        "type": "object"
      },
...
      "ProcessSingleRequest": {
        "properties": {
          "image_path": {
            "title": "Image Path",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/ProcessingOptions"
          }
        },
        "required": [
          "image_path",
          "options"
        ],
        "title": "ProcessSingleRequest",
        "type": "object"
      },
      "ProcessingOptions": {
        "properties": {
          "alpha_threshold": {
            "default": 10,
            "title": "Alpha Threshold",
            "type": "integer"
          },
          "aspect_ratio": {
            "default": "keep",
            "title": "Aspect Ratio",
            "type": "string"
          },
          "background_color": {
            "default": "#FFFFFF",
            "title": "Background Color",
            "type": "string"
          },
          "background_mode": {
            "default": "transparent",
            "enum": [
              "transparent",
              "solid"
            ],
            "title": "Background Mode",
            "type": "string"
          },
          "cutout_engine": {
            "const": "rembg",
            "default": "rembg",
            "title": "Cutout Engine",
            "type": "string"
          },
          "cutout_model_id": {
            "default": "u2netp",
            "title": "Cutout Model Id",
            "type": "string"
          },
          "edge_feather_radius": {
            "default": 1,
            "title": "Edge Feather Radius",
            "type": "integer"
          },
          "enhance_engine": {
            "const": "realesrgan",
            "default": "realesrgan",
            "title": "Enhance Engine",
            "type": "string"
          },
          "enhance_level": {
            "default": "off",
            "enum": [
              "off",
              "2x",
              "4x"
            ],
            "title": "Enhance Level",
            "type": "string"
          },
          "enhance_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enhance Model"
          },
          "fallback_to_api": {
            "default": false,
            "title": "Fallback To Api",
            "type": "boolean"
          },
          "filename_pattern": {
            "default": "{original_name}_{preset}_{engine}",
            "title": "Filename Pattern",
            "type": "string"
          },
          "local_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Model"
          },
          "local_quality_preset": {
            "anyOf": [
              {
                "enum": [
                  "fast",
                  "balanced",
                  "hq"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Quality Preset"
          },
          "naming_mode": {
            "default": "pattern",
            "enum": [
              "keep_original",
              "pattern",
              "ocr_text"
            ],
            "title": "Naming Mode",
            "type": "string"
          },
          "naming_regex_find": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Naming Regex Find"
          },
          "naming_regex_replace": {
            "default": "",
            "title": "Naming Regex Replace",
            "type": "string"
          },
          "ocr_language": {
            "default": "eng",
            "title": "Ocr Language",
            "type": "string"
          },
          "ocr_max_length": {
            "default": 48,
            "title": "Ocr Max Length",
            "type": "integer"
          },
          "output_dir_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Dir Override"
          },
          "output_format": {
            "default": "png",
            "enum": [
              "png",
              "webp",
              "jpeg",
              "jpg",
              "avif"
            ],
            "title": "Output Format",
            "type": "string"
          },
          "padding": {
            "default": 0,
            "title": "Padding",
            "type": "integer"
          },
          "preprocess_color_normalization": {
            "default": false,
            "title": "Preprocess Color Normalization",
            "type": "boolean"
          },
          "preprocess_denoise": {
            "default": false,
            "title": "Preprocess Denoise",
            "type": "boolean"
          },
          "preprocess_sharpening": {
            "default": false,
            "title": "Preprocess Sharpening",
            "type": "boolean"
          },
          "preset": {
            "default": "quick_cutout",
            "title": "Preset",
            "type": "string"
          },
          "processing_order": {
            "anyOf": [
              {
                "enum": [
                  "cutout_then_enhance",
                  "enhance_then_cutout"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "cutout_then_enhance",
            "title": "Processing Order"
          },
          "provider_priority": {
            "items": {
              "type": "string"
            },
            "title": "Provider Priority",
            "type": "array"
          },
          "quality": {
            "default": 90,
            "title": "Quality",
            "type": "integer"
          },
          "remove_background": {
            "default": true,
            "title": "Remove Background",
            "type": "boolean"
          },
          "resize_max_height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resize Max Height"
          },
          "resize_max_width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resize Max Width"
          },
          "resize_mode": {
            "default": "keep",
            "enum": [
              "keep",
              "custom"
            ],
            "title": "Resize Mode",
            "type": "string"
          },
          "save_alpha_mask": {
            "default": false,
            "title": "Save Alpha Mask",
            "type": "boolean"
          },
          "strip_metadata": {
            "default": true,
            "title": "Strip Metadata",
            "type": "boolean"
          },
          "trim_transparent_bounds": {
            "default": true,
            "title": "Trim Transparent Bounds",
            "type": "boolean"
          },
          "white_halo_cleanup": {
            "default": 35,
            "title": "White Halo Cleanup",
            "type": "integer"
          },
          "workflow_mode": {
            "default": "cutout_only",
            "enum": [
              "cutout_only",
              "enhance_only",
              "cutout_enhance"
            ],
            "title": "Workflow Mode",
            "type": "string"
          }
        },
        "title": "ProcessingOptions",
        "type": "object"
      },
...
      "UploadItem": {
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          },
          "storage_mode": {
            "default": "uploaded_blob",
            "enum": [
              "desktop_path",
              "uploaded_blob"
            ],
            "title": "Storage Mode",
            "type": "string"
          },
          "upload_id": {
            "title": "Upload Id",
            "type": "string"
          }
        },
        "required": [
          "upload_id",
          "filename",
          "size",
          "mime_type",
          "path"
        ],
        "title": "UploadItem",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ClearCut",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "post": {
        "operationId": "create_job_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Job Jobs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/batch": {
      "post": {
        "operationId": "create_batch_job_jobs_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Batch Job Jobs Batch Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Batch Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/download": {
      "get": {
        "operationId": "download_output_jobs_download_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Output",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/ingest-paths": {
      "post": {
        "operationId": "ingest_local_paths_jobs_ingest_paths_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestPathsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UploadItem"
                  },
                  "title": "Response Ingest Local Paths Jobs Ingest Paths Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Local Paths",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/single": {
      "post": {
        "operationId": "process_single_jobs_single_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessSingleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Process Single Jobs Single Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Process Single",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/upload": {
      "post": {
        "operationId": "upload_files_jobs_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_files_jobs_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UploadItem"
                  },
                  "title": "Response Upload Files Jobs Upload Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Files",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_status_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Status",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Job Jobs  Job Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/zip": {
      "get": {
        "operationId": "download_job_zip_jobs__job_id__zip_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Job Zip",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}